libsysbase: Add libsysbase from devkitPro
This commit attempts to pull machine independent pieces from
libsysbase from the following files:
  buildscripts/dkarm-eabi/patches/newlib-1.20.0.patch
  buildscripts/dkppc/patches/newlib-1.20.0.patch

Those patches were taken from devkitPro's buildscripts-20120625.

+++ newlib/libc/stdio/fseeko.c
@@ -26,7 +26,7 @@ _DEFUN(_fseeko_r, (ptr, fp, offset, whence),
        _off_t offset      _AND
        int whence)
 {
-  return _fseek_r (ptr, fp, (long)offset, whence);
+  return _fseek_r (ptr, fp, offset, whence);
 }
 
 #ifndef _REENT_ONLY
@@ -38,7 +38,7 @@ _DEFUN(fseeko, (fp, offset, whence),
        int whence)
 {
   /* for now we simply cast since off_t should be long */
-  return _fseek_r (_REENT, fp, (long)offset, whence);
+  return _fseek_r (_REENT, fp, offset, whence);
 }
 
 #endif /* !_REENT_ONLY */
